+++ tests/test_config.py
@@ -42,3 +42,26 @@ def test_config_file_loading(tmp_path, caplog: pytest.LogCaptureFixture):
     create_app(test_config=None, instance_path=tmp_path)
     assert "Using this path as it's the first one that was found" in caplog.text
     caplog.clear()
+
+def test_config_invalid(tmp_path, get_test_config):
+    """Test that program exits when given invalid config."""
+    # TEST: Assert that the program exists when provided an invalid config dictionary.
+    with pytest.raises(SystemExit) as exc_info:
+        create_app(test_config=get_test_config("invalid"), instance_path=tmp_path)
+
+    assert isinstance(exc_info.type, type(SystemExit)), "App did not exit on config validation failure."
+    assert exc_info.value.code == 1, "App did not have correct exit code for config validation failure."
+    os.unlink(os.path.join(tmp_path, "config.toml"))
+
+
+def test_config_file_creation(tmp_path, caplog: pytest.LogCaptureFixture):
+    """Tests relating to config file."""
+    # TEST: that file is created when no config is provided.
+    import contextlib
+    with contextlib.suppress(FileNotFoundError):
+        os.unlink(os.path.join(tmp_path, "config.toml"))
+
+    with caplog.at_level(logging.WARNING):
+        create_app(test_config=None, instance_path=tmp_path)
+
+    assert "No configuration file found, creating at default location:" in caplog.text
